osd: make object_locator_t encodable

+++ src/osdc/ObjectCacher.h
@@ -297,7 +297,7 @@ class ObjectCacher {
   Object *get_object_maybe(sobject_t oid, object_locator_t &l) {
     // have it?
     assert(l.pool >= 0);
-    if ((l.pool < (int)objects.size()) &&
+    if ((l.pool < objects.size()) &&
         (objects[l.pool].count(oid)))
       return objects[l.pool][oid];
     return NULL;
@@ -306,7 +306,7 @@ class ObjectCacher {
   Object *get_object(sobject_t oid, ObjectSet *oset, object_locator_t &l) {
     // have it?
     assert(l.pool >= 0);
-    if (l.pool < (int)objects.size()) {
+    if (l.pool < objects.size()) {
       if (objects[l.pool].count(oid))
         return objects[l.pool][oid];
     } else {
